Update PR. Add new source and header files, touched up the logger logic.
